refactor: split llm.rs into LlmProvider trait + per-provider modules (issue #61)
Introduce the LlmProvider async trait to replace five scattered `match provider { ... }`
blocks. Each provider owns its HTTP client, SSE parsing, usage extraction, streaming
format, and error messages. Adding a new provider now requires one file in
`src/llm/providers/` and one line in `get_provider()`.

Changes:
- Add `async-trait` dependency
- Extract API key management to `src/keystore.rs`
- Split `llm.rs` (3649 lines) into focused modules:
  - `llm/types.rs` — domain structs (GlossaryEntry, PipelineConfig, etc.)
  - `llm/provider.rs` — LlmProvider trait, LlmRequest/LlmResponse, UsageAccumulator
  - `llm/stream.rs` — StreamRegistry, CancelToken, consume_stream, StreamAccumulator
  - `llm/prompts.rs` — prompt builders, sanitize_llm_json_output, parse_judge_rating
  - `llm/pipeline.rs` — Tauri commands (run_stage, judge_translation, etc.)
  - `llm/providers/{gemini,openai,anthropic,ollama}.rs` — one file per backend
- Eliminate call_provider_for_judge and build_streaming_request match blocks entirely
- StreamAccumulator now calls provider.extract_streaming_token() and
  provider.update_streaming_usage() instead of matching on provider id strings
- All 75 existing tests pass unchanged
