+++ src-tauri/src/llm/types.rs
@@ -1,0 +1,107 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct GlossaryEntry {
+    pub term: String,
+    pub translation: String,
+    pub notes: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OllamaConfig {
+    pub temperature: Option<f32>,
+    pub top_p: Option<f32>,
+    pub seed: Option<i32>,
+    pub keep_alive: Option<serde_json::Value>,
+    pub think: Option<serde_json::Value>,
+    pub num_ctx: Option<u32>,
+    pub num_predict: Option<i32>,
+    pub use_advanced_options: Option<bool>,
+    pub advanced_options: Option<serde_json::Map<String, serde_json::Value>>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct ProviderRuntimeConfig {
+    pub ollama: Option<OllamaConfig>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct StageConfig {
+    pub id: String,
+    pub name: String,
+    pub prompt: String,
+    pub model: String,
+    pub provider: String,
+    pub enabled: bool,
+    pub provider_options: Option<ProviderRuntimeConfig>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct PipelineConfig {
+    pub source_language: String,
+    pub target_language: String,
+    pub stages: Vec<StageConfig>,
+    pub judge_prompt: String,
+    pub judge_model: String,
+    pub judge_provider: String,
+    pub glossary: Vec<GlossaryEntry>,
+    pub use_chunking: Option<bool>,
+    pub markdown_aware: Option<bool>,
+    pub coherence_prompt: Option<String>,
+    pub review_provider_options: Option<ProviderRuntimeConfig>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct OllamaPreflightStatus {
+    pub reachable: bool,
+    pub models: Vec<String>,
+    pub requested_model: Option<String>,
+    pub model_available: bool,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct JudgeIssue {
+    #[serde(rename = "type")]
+    pub issue_type: String,
+    pub severity: String,
+    pub description: String,
+    pub suggested_fix: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct JudgeResponse {
+    pub rating: String,
+    pub issues: Vec<JudgeIssue>,
+    pub content: String,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub input_tokens: Option<u32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub output_tokens: Option<u32>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CoherenceChunkInput {
+    pub original: String,
+    pub translation: String,
+    pub prev_context: Option<String>,
+    pub next_context: Option<String>,
+}
+
+#[derive(Debug, Clone, Serialize, Deserialize)]
+#[serde(rename_all = "camelCase")]
+pub struct CoherenceResponse {
+    pub issues: Vec<JudgeIssue>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub input_tokens: Option<u32>,
+    #[serde(skip_serializing_if = "Option::is_none")]
+    pub output_tokens: Option<u32>,
+}
